Oragnize GameFunction validation methods to GameFunctionValidator class
